refac(zk): move `rotation` to `zk/base`
