+++ tachyon/zk/plonk/constraint_system/query.h
@@ -11,8 +11,8 @@
 
 #include "absl/strings/substitute.h"
 
+#include "tachyon/zk/base/rotation.h"
 #include "tachyon/zk/plonk/base/column_key.h"
-#include "tachyon/zk/plonk/constraint_system/rotation.h"
 
 namespace tachyon::zk {
