+++ tachyon/zk/plonk/constraint_system/virtual_cell.h
@@ -4,8 +4,8 @@
 #include <string>
 
 #include "tachyon/export.h"
+#include "tachyon/zk/base/rotation.h"
 #include "tachyon/zk/plonk/base/column_key.h"
-#include "tachyon/zk/plonk/constraint_system/rotation.h"
 
 namespace tachyon::zk {
