+++ tachyon/zk/plonk/permutation/BUILD.bazel
@@ -59,8 +59,8 @@ tachyon_cc_library(
         "//tachyon/base:ref",
         "//tachyon/crypto/commitments:polynomial_openings",
         "//tachyon/zk/base:point_set",
+        "//tachyon/zk/base:rotation",
         "//tachyon/zk/base/entities:prover_base",
-        "//tachyon/zk/plonk/constraint_system:rotation",
     ],
 )
