+++ tachyon/zk/plonk/vanishing/BUILD.bazel
@@ -23,11 +23,11 @@ tachyon_cc_library(
         "//tachyon/base:parallelize",
         "//tachyon/base/containers:container_util",
         "//tachyon/base/numerics:checked_math",
+        "//tachyon/zk/base:rotation",
         "//tachyon/zk/lookup:lookup_committed",
         "//tachyon/zk/plonk/base:column_key",
         "//tachyon/zk/plonk/base:owned_table",
         "//tachyon/zk/plonk/base:ref_table",
-        "//tachyon/zk/plonk/constraint_system:rotation",
         "//tachyon/zk/plonk/keys:proving_key_forward",
         "//tachyon/zk/plonk/permutation:permutation_committed",
         "//tachyon/zk/plonk/permutation:unpermuted_table",
