+++ tachyon/zk/plonk/vanishing/circuit_polynomial_builder.h
@@ -14,11 +14,11 @@
 #include "tachyon/base/containers/container_util.h"
 #include "tachyon/base/numerics/checked_math.h"
 #include "tachyon/base/parallelize.h"
+#include "tachyon/zk/base/rotation.h"
 #include "tachyon/zk/lookup/lookup_committed.h"
 #include "tachyon/zk/plonk/base/column_key.h"
 #include "tachyon/zk/plonk/base/owned_table.h"
 #include "tachyon/zk/plonk/base/ref_table.h"
-#include "tachyon/zk/plonk/constraint_system/rotation.h"
 #include "tachyon/zk/plonk/keys/proving_key_forward.h"
 #include "tachyon/zk/plonk/permutation/permutation_committed.h"
 #include "tachyon/zk/plonk/permutation/unpermuted_table.h"
